requirements.txt
scrapy>=2.11
SQLAlchemy>=2.0
pandas>=2.2
openpyxl>=3.1
python-dateutil>=2.9
psycopg[binary]>=3.2
